requirements.txt
numpy>=1.23.5,<2
typing-extensions>=4.8.0
opencv-python==4.10.0.84
cv2_enumerate_cameras==1.1.15
onnx==1.18.0
insightface==0.7.3
psutil==5.9.8
customtkinter==5.2.2
pillow==12.1.1
tqdm>=4.66.0
scikit-learn>=1.3.0

# PyTorch
torch==2.10.0; sys_platform != 'darwin'
torchvision==0.25.0; sys_platform != 'darwin'
torch==2.8.0; sys_platform == 'darwin'
torchvision==0.20.1; sys_platform == 'darwin'

# ONNX Runtime
onnxruntime-gpu==1.22.0; sys_platform != 'darwin'
onnxruntime-silicon==1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'

# NSFW / auxiliary runtime
tensorflow==2.19.1; sys_platform != 'darwin'
opennsfw2==0.10.2
protobuf==4.25.1

# Face restoration stack
basicsr==1.4.2
gfpgan==1.3.8
facexlib>=0.3.0
realesrgan>=0.3.0
filterpy>=1.4.5
pyyaml>=6.0.0
requests>=2.31.0

# Platform-specific capture helpers
pygrabber; sys_platform == 'win32'
